--- a/crates/compute_test_utils/Cargo.toml
+++ b/crates/compute_test_utils/Cargo.toml
@@ -18,4 +18,6 @@ binius_math = { path = "../math", default-features = false }
 binius_ntt = { path = "../ntt", default-features = false }
 binius_utils = { path = "../utils", default-features = false }
 bytemuck = { workspace = true, features = ["extern_crate_alloc"] }
+itertools.workspace = true
 rand = { workspace = true, features = ["std"] }
+