[piop] Fix failing PIOP case when there are no evaluation claims (#221)

Previously, when a committed column had no evaluation claims, the PIOP compiler would throw a Math(PiecewiseMultilinearIncompatibleEvals error. While an unconstrained committed is bad, this should get warned about at a higher level, not throw a confusing error deep in the stack.

[piop] Handle arbitrarily small committed polynomials (#227)

This relaxes the ring_switch and piop modules to handle committed polynomials that are smaller than the size of a packed large element and eliminates the OracleTooSmall error. We handle this case by committing the polynomial as it's MLE over the larger hypercube, which corresponds to repeating it's evaluations. The integration test in ring_switch module is updated to exercise this code path.

[m3]: add test for U32Add gadget (#229)

Add fixing generics for `DeserializeBytes` derive macro (#212)

This PR fixes the issue #88 by adding a custom attribute eval_generics for the derive macro.

Change MultilinearExtension::from_vales to MultilinearExtension::new (#234)

[m3] chore add u32 mod (#233)

Creates a module in anticipation of other 32-bit ALU gadgets, e.g. U32Sub.

[m3] Fix the small-table flushing bug for lookups (#230)

This issue is caused by the solution of #170, which makes it possible for tables to be automatically allocated capacity exceeding their event count. That works fine for regular tables, but not lookup tables that are power_of_two_sized. However, that solution is now obsolete because of an alternative fix in #227, so this adds a test and assertion and rolls the change back.

[fri] Fix shape of terminal FRI oracle (#231)

Update the batch_size of the terminate_codeword Merkle tree so that it has exactly inv_rate leaves.

[m3] fix OOB in Display impl of ConstraintSystem (#237)

[math] Fast (single pass) algorithm for `ArithExpr::linear_normal_form` (#224)

[evalcheck] Add documentation for Evalcheck (#226)

[m3] u32 sub gadget (#242)

[ntt] Remove overly conservative domain size check in AdditiveNTT (#236)

Always sample 4 bytes during `sample_bits` (#225)

[circuits] Optimised Blake3 compression (#222)

[m3] u32 gadgets follow-up (#243)

1. Tidy the lingering dbg! in u32sub
2. Move u32add to the same test approach as u32sub

[serialization] Serialize usize as u32 instead of u64 (#245)

[compute] Interface for host memory allocation (#240)

Improve instrumentation (#238)

fix: inf-178: fix nightly benchmarks (#246)

[math] Common expression extraction in `ArithExpr` (#235)

[evalcheck] fix ZeroPadded (#247)

SYS-110:: define ComputeAllocator trait (#248)

Accelerated compute backends will provide custom implementations for
host and device allocations. Define a trait for bump allocators that
will be used throughout Binius for compute buffer allocations.

[M3] Add barrel shifter gadget (#249)

[gadgets] Adds unsigned multiplication gadgets to M3 (#239)

This MR adds unsigned multiplication gadgets to M3, as well as expose GKR exponentiation into M3.

[toolchain] nightly-2025-04-24 (#252)

bump toolchain version

[m3] Revert the #230 related to the small lookup table (#256)

This change caused issues that we are still investigating.

[deps] Bump tracing-profile version (#251)

ci: INF-176 oom fix (#264)

[zerocheck] Improve univariate skip (#244)

Fixes the zerocheck univariate round soundness hole by always skipping the same number of rounds, resulting in same-sized round polynomials in Lagrange basis. Remaining rounds are "right justified" using front-loaded batching, resulting in a peculiar reduction shape where projection happens on the low variables (that underwent "univariate skip") and possibly a suffix of high variables, potentially leaving a "gap" in between. The primary reason for this contrived arrangement is to be able to reuse high projection from univariatizing reduction witness generation in ring switching.

[m3] Port keccak-f to M3 (#259)

[m3] Test utility to simplify testing systems & gadgets (#262)

validate_system_witness by default just does naive witness validation, but an environment variable can be toggled to perform proving + proof verification.

[examples] Keccak-f example using M3 (#260)

[oracle] Generalize ZeroPadding (#255)

Generalize the ZeroPadded virtual oracle with

- num_extra_variables: number of additional variables.
- start_index: variable index (within the set of all variables) where the padding starts
- nonzero_index: the index (between 0 and (1 << num_extra variables) - 1) of the block (of size 1 << start_index) where the values are those of the original column.

Move tower from core to field (#272)

[M3] Fix single-element tables (#271)

Fixes evalcheck and gkr_gpa to support poly with n_vars=0

[m3] Public method for ZeroPadding with nonzero_index as an argument (#273)

Public method with nonzero_index as an argument

[m3] Add structured column type (#269)

This adds dynamically-sized structured columns. In a future change, we will add fixed-size structured columns as well.

[compute] Extend with functionality for multilinear evaluation (#266)

Extends ComputeLayer with test suite (generalised over compute layer) for some multilinear evaluation and needed functionality in the layer trait itself.

[reed_solomon] Extract NTT from ReedSolomonCode (#274)

[channels] Handle selector logic with CompositeMLE (#270)

Change the manual handling of flushes with selectors into evalcheck using composite MLE.

[chore] tidy keccakf example (#279)

- removes the old keccakf_circuit.rs,
- uses the new keccakf M3 port as the nightly benchmark target,
- cleans up the new keccak wrt seq/par table filling.

[m3] use only referenced oracle ids for zero check (#278)

Instead of unconditionally adding the whole list of per partition oracle ids into the constraint set, we check which ones are actually getting used.

Fixes #265

[evalcheck] Write duplicate claim advice directly to transcript (#277)

This removes most of the serialization/deserialization & proof struct pattern matching logic from evalcheck.

[m3] Fixed-size structured columns (#280)

[field] please clippy (#283)

This one is the same change as in #257 but isolated for faster merging.

[m3] fix rustdoc for the keccak module (#284)

[m3] add TableBuilder doc focusing on nesting (#285)

Also actually add a test for checking if the nesting works properly.

[fri] Simplify the FRI folding code (#282)

This simplifies the FRI folding code in a few ways. The improvements are:

* Remove the requirement that fold_challenges be non-empty
* Remove the fold_codeword prover function and use the more general fold_interleaved in all cases
* Pass log_len instead of round parameter so that the code will work with NTT instances with larger domains

Fix zero pad check and improve error message (#288)

[arith_expr] Use `ArithExpr` as a builder only (#250)

All other operations are now moved to `ArithCircuit`.

[m3] namespace zero constraints (#286)

This is important because otherwise instantiations of the same gadget will have all the constraints having the same name.

[gadgets] Signed Multiplication Gadgets (#257)

* This MR adds MulSU32, MulSS32, TwosComplement, Incr, SignConverter gadgets.
* Added better testing which can possibly be reused in other gadgets.

[m3] display assert_zero costs per table. (#289)

This changeset aims to give more insight for the approximate cost of a given table.

The cost of assert_zero constraints is:

(Degree - 1) x (# of muls in expr) x V.

Then that is multiplied by the bits in the field (with 8-bit as floor).

[m3] Stat returns bits per row (#292)

[evalcheck] Handle composite MLEs separately from projected bivariate sumchecks (#290)

They are different and should be handled differently. Composite MLE evaluations are not bivariate, and they don't use projection. In the future, we shouldn't handle them as regular sumchecks, we should prove them as MLEchecks, which have a more efficient proving algorithm.

[compute] Reduce duplication in CpuLayer using macro (#295)

[field] Optimized byte sliced `spread` implementation (#276)

[examples] Fix the bitwise XOR example (#296)

[m3] emulate Channel relax Default constraint (#298)

Channel prior to this PR is overconstrained. It requires T to implement Default all the while the Channel does not actually contain any T by default.

[M3] Allow core constraint system to handle flushes with multiple selectors (#294)

[Fix] Match the order of processing bivariate sumcheck claims in evalcheck (#297)

The verifier processes claims successively, while the prover processes all projected claims first, followed by all composite. As a result, if a projected and a composite claim have the same n_vars, the order of sumcheck_claims will differ between the verifier and prover.

Replace backloaded with frontloaded sumchecks (#281)

Replaces some backloaded sumchecks with frontloaded suchecks, towards keeping only frontloaded sumchecks. The goal is to simplify and reduce code.

[M3] Automatic witness population for constants (#304)

Small MR to automatically populate with witness index for constants.

[m3] Surface error when lookup count is too high (#302)

[compute] Define HAL operation for fold_left (#293)

[ntt] Add skip_rounds parameter to AdditiveNTT (#306)



This generalizes the AdditiveNTT operation to be able to perform a truncated NTT, which skips rounds at the beginning of a forward transform and the end of an inverse transform. This is useful for Reed-Solomon encoding, where we want to run a large NTT but know apriori that the first few rounds have the effect of simply copying data.

This resolves a limitation of ReedSolomonCode::encode, which imposes a packing length restriction on the encoded message and also complicates the multithreading around the NTT with an additional Rayon loop. Now the multithreading is delegated to the AdditiveNTT implementation, which can size the matrix more appropriately.

[m3] Fix bug in filling structured columns (#305)

[evalcheck] Fix ZeroPadded memoization (#308)

[gadgets] Division gadgets for M3 (#258)

[m3] emulate::Channel::assert_balanced (#309)

This allows to actually print what values are unbalanced which could aid debugging.

[compute] Introduce `accumulate_kernels` (#299)

[gkr_gpa] Switch to front-loaded batch sumcheck (#312)

[m3] destruct gadgets::u32 module (#311)

As you may see already in the u32::add and u32::sub modules, those are no longer
32-bit only but support 64-bit ops as well.

As was discussed elsewhere categorizing gadgets based on operation and treating
the width as that op's parameter is the way forward.

[arith_expr] Fix `pow` step evaluation in `ArithCircuit` (#315)

[tracing] Add data dimensions tracing (#263)

[utils] remove unsound thread_local_mut (#317)

The thread_local_mut utility is unsound as it can provide multiple exclusive
references to the inner value as shown by the test below:

	#[test]
	fn evil() {
		let tlm = super::ThreadLocalMut::new();

		tlm.with_mut(
			|| vec![1, 2, 3],
			|exclusive_1| {
				tlm.with_mut(
					|| vec![],
					|exclusive_2| {
						exclusive_1.push(1);
						exclusive_2.push(1);
					},
				)
			},
		);
	}

Since it seemingly is not used anywhere, removing it.

[bug] Fix bug with exponentiation constraints on small tables (#313)

[ci] Run M3 integration tests in CI (#320)

[compute] Add FRI fold operation to ComputeLayer (#314)

[evalcheck] Serialize evalcheck proofs using u32s instead of bytes and usize (#323)

[m3] High-level model for Grøstl-256 hash function (#310)

Implementation of the high-level model of Grøstl-256. The presented version is specialized to a fixed length input size: 1024 bytes. This is done because otherwise it would be harder to arithmetize the logic of adding padding. It would be easy to change the length of the input as long as it is a multiple of Grøstl-256 block size (64 byte).

[m3] eval_expr bug fix (#321)

[core] FRI fold include num_challenges in dimensions trace (#319)

[fri] Improve early FRI termination optimization (#300)

[fri] Add FRI folding tower level logging (#328)

[m3] Test coverage & cleanups for multiplication gadget (#318)

Add unit test to catch issue with compilation bug. Also test cleanups in mul gadget.

[m3] enforce the fixed power-of-two table size constraint (#324)

[m3] log_capacity does not depend on &self (#325)

Small refactor and docs update to avoid confusion.

Why I didn't just inline it? Well then there won't be a platform for the
documentation.

[rustfmt] Enforce maximum comment width (#326)

[Fix] gkr_exp claims descending order (#330)

Descending claim order was disrupted when dynamic claims began being processed after static ones — this PR restores the correct order.

[core] newtype OracleId (#327)

OracleId becomes a newtype. This is a non-functional change.

As a bonus: turns out there were quite a bit of clones of MultilinearPolyOracle
struct. Each one is, if instantiated with e.g. F=BinaryField32b is 160 bytes.
Not too much but definitely annoying, especially considering why it was
happening. In one instance the whole clone of the structure would be done
just to get the name/label.

Update github CI pull_request types to support graphite (#335)

The graphite docs recommend this change to the github action `on` spec:

https://graphite.dev/docs/github-configuration-guidelines#github-actions

This is because PRs are stacked and you want to run CI tests on all PRs
in the stack, not just those that are based on the main.

The `edited` type is excluded here because this event is triggered when
Graphite makes changes to PR metadata and we don't want this to trigger
a CI run.

[compute] Fold right HAL operation (#334)

[evalcheck] Reduce CompositeMLE evaluations to MLEchecks (#307)

Use EqIndSumcheckProver (MLECheck) to prove CompositeMLE oracles; deduplication is driven by advice tape.

[m3] Exposed input and output columns for Keccak-f[1600] gadget. (#329)

Updated the keccak gadget to:
* Expose the input and output columns to the gadget.
* Change the interface for the constructor and populate methods.

[ci]: Allow manual trigger of benchmark job (#340)

[compute] Add `kernel_add` operation (#336)

[compute] Simplify reference fri_fold operation (#333)

The fold interleaved step is invariant to the selected batch size. Refactor the reference CPU implementation to make this property clear. Hardware implementations will benefit from the ability to select different batch sizes based on the properties of the target device.

[m3] Indexed lookups with tallying (#331)



Indexed lookup tables are fixed-size tables where every entry is easily determined by its index.

Indexed lookup tables cover a large and useful class of tables, such as lookup tables for bitwise operations, addition of small integer values, multiplication, etc. The entry encodes an input and an output, where the index encodes the input. For example, a bitwise AND table would have 2 8-bit input values and one 8-bit output value. The index encodes the input by concatenating the 8-bit inputs into a 16-bit unsigned integer.

Before a lookup table witness can be filled, the number of times each entry is read must be known. Reads from indexed lookup tables are a special case where the counts are difficult to track during emulation, because the use of the lookup tables is an arithmetization detail. For example, emulation of the system model should not need to know whether integer additions within a table are constraint using zero constraints or a lookup table for the limbs. In most cases of practical interest, the lookup table is indexed.

The method to tally counts is to scan all tables in the constraint system and boundaries values, and identify those that pull from the lookup table's channel. Then we iterate over the values read from the table and count all the indices.

[m3] fix underconstraining bug in U32Sub (#349)

Closes [CRY-347].

[CRY-347]: https://linear.app/irreducible/issue/CRY-347/m3-u32sub-gadgets-doesnt-constrain-borrow-out

[gkr_gpa] Change sumcheck order to high-to-low (#339)

Changed the evaluation order in grand product argument from `LowToHigh` to `HighToLow`.

[deps]: bump `bytesize` and `hex-literal` (#347)

[CODEOWNERS] Add Dmytro as codeowner (#350)

[examples] Use fill_table_parallel instead of sequential (#344)

This will speed up witness generation for examples in multithreaded configurations.

[rust] Update from Rust 2021 to Rust 2024 edition (#351)

Most of the file changes are due to rustfmt sorting dependencies differently in Rust 2024 edition.
- `rng.gen()` had to be changed to `rng.r#gen()` since gen is now a reserved keyword in Rust 2024.
- All unsafe functions must now wrap any unsafe code it calls in explicit unsafe blocks

[m3] Update indexed_lookup incr example with column permutation (#332)

This makes the incr lookup table example more realistic by adding the column permutation. The only change to the library code is the addition of an index_to_entry method on IndexedLookup.

[ring_switch] Ring-Switch EQ Indicator using evaluate_partial_low (#348)

Using evaluate_partial_low (a HAL operation) instead of the current inner_product_subfield to calculate the coefficients of the MLE of the ring-switch EQ indicator .

[utils] remove mod examples (#352)

It does not seem that this code is used anywhere.

[m3] Optimize the LookupGadget using the special truncated sumcheck (#343)

Added automatic zero suffix detection for CompositeMLE in evalcheck and optimized make_masked_flush_witnesses

[m3] slap track_caller on assert_balanced (#364)

Add `actual` to every query size error (#366)

This is mostly a mechanical change.

[m3] Document some TableBuilder methods (#360)

INF-194: Run cargo docs tests (#365)

[utils] introduce mem::slice_uninit_mut (#355)

As per @GraDKh request in this [comment].

I did not use the suggested identifier as, unlike in `assume_init`, we don't
really assume anything here, we are just converting a definitely inited slice
into a maybe uninited slice. Also to make it inline with the follow-up I
decided to call it `slice_uninit_mut` (put slice in the front).

[comment]: https://app.graphite.dev/github/pr/IrreducibleOSS/binius/353/%5Bcore%5D-miri

[utils] prefer homegrown slice_assume_init (#356)

We already have slice_assume_init at home.

[m3] Binary Merkle Tree (#362)

This PR adds a high-level model for binary Merkle trees using the Grøstl-256 output transformation as a 2-to-1 compression function.

[pre-commit] Fix rustfmt pre-commit hook (#372)

Use cargo fmt instead of rustfmt for pre-commit hook

[zerocheck] Step towards reusing ntt (#373)

This make a step towards reusing the already-created NTT object in extrapolate_round_evals. We still can't actually use it because of extension field nonsense, but at least this ensures that the NTT domains are consistent.

[m3] don't assume oracle id order in witness (#357)

This changeset aims to fix the implicit dependency of the witness on the order
of oracles added in `ConstraintSystem::compile`.. See [CRY-356] for more
details.

This does several things:

1. Prefers using ColumnId over indices
2. Introduction of `OracleLookup`

This starts using the ColumnId where possible. This change should really make
much less likely to confuse indices and improve general type-safety, and as well
as simplify the code a bit (one type is better than multiple). However, this
makes the column IDs to carry table_id around, but I don't think this is a
problem since the code is not hot.

Introduction the `OracleLookup` type. Prior this changeset
the data about which oracle came from which columns was lost. This type
is meant to store this information explicitly to be read by the witness
handling code.

This opens up for other use-cases where you might want to pass some data
from the compilation process into the witness generation, as that seems to be
inherently important: compilation creates oracles from columns and witness
…